+++ lib/sbb_design_system_mobile.dart
@@ -30,6 +30,7 @@ export 'src/radio_button/sbb_radio_button.dart';
 export 'src/radio_button/sbb_radio_button_list_item.dart';
 export 'src/segmented_button/sbb_segmented_button.dart';
 export 'src/select/sbb_select.dart';
+export 'src/radio/radio.dart';
 export 'src/shared/closeable_box_controller.dart';
 export 'src/slider/sbb_slider.dart';
 export 'src/status/sbb_status.dart';