+++ lib/src/radio/sbb_radio_list_item.dart
@@ -1,27 +1,27 @@
 import 'package:flutter/material.dart';
+import 'package:sbb_design_system_mobile/src/checkbox/bottom_loading_indicator.dart';
 
 import '../../sbb_design_system_mobile.dart';
-import 'bottom_loading_indicator.dart';
 
 const double _iconTopPadding = 10.0;
 const double _minListTileHeight = 44.0;
 
-/// SBB Checkbox Item. Use according to documentation.
+/// SBB Radio Button List Item. Use according to documentation.
 ///
-/// Use [SBBControlStyle].listItem to manipulate the style of this widget.
+/// Use [SBBControlStyle].listItem to manipulate the style of this Widget.
 ///
 /// See also:
 ///
-/// * [SBBCheckbox], which is a part of this widget.
-/// * [SBBSwitch], a widget with semantics similar to [SBBCheckbox].
-/// * [SBBRadioListItem] and [SBBRadio], for selecting among a set
-/// of explicit values.
+/// * [SBBRadio], which is a part of this widget.
+/// * [SBBSegmentedButton], a widget with semantics similar to [SBBRadio].
 /// * [SBBSlider], for selecting a value in a range.
-/// * <https://digital.sbb.ch/de/design-system/mobile/components/checkbox/>
-class SBBCheckboxListItem extends StatelessWidget {
-  /// Creates a combination of a tile and a [SBBCheckbox].
+/// * [SBBCheckboxListItem], [SBBCheckbox] and [SBBSwitch], for toggling a
+/// particular value on or off.
+/// * <https://digital.sbb.ch/de/design-system-mobile-new/elemente/radiobutton>
+class SBBRadioListItem<T> extends StatelessWidget {
+  /// Creates a combination of a tile and a [SBBRadio].
   ///
-  /// The [SBBCheckboxListItem] itself does not maintain any state. Instead, when the
+  /// The [SBBRadioListItem] itself does not maintain any state. Instead, when the
   /// state of the checkbox changes, the widget calls the [onChanged] callback.
   /// Most widgets that use a checkbox will listen for the [onChanged] callback
   /// and rebuild the checkbox tile with a new [value] to update the visual
@@ -29,11 +29,10 @@ class SBBCheckboxListItem extends StatelessWidget {
   ///
   /// The following arguments are required:
   ///
-  /// * [value], which determines whether the checkbox is checked. The [value]
-  ///   can only be null if [tristate] is true.
-  /// * [label], the primary text written on [SBBCheckboxListItem].
-  /// * [onChanged], which is called when the value of the checkbox should
-  ///   change. It can be set to null to disable the checkbox.
+  /// * [value] and [groupValue] together determine whether the radio button is
+  ///   selected.
+  /// * [label], the primary text written on [SBBRadioListItem].
+  /// * [onChanged] is called when the user selects this radio button.
   ///
   /// Set the [isLastElement] true for the last item in a list to not show any Divider.
   ///
@@ -43,149 +42,134 @@ class SBBCheckboxListItem extends StatelessWidget {
   /// * if [onCallToAction] is not given, the trailingIcon will be shown as 24px sized [Icon].
   /// * if [onCallToAction] is given, the trailingIcon will be wrapped in a [SBBIconButtonSmall].
   ///
-  ///
   /// If [isLoading] is true, a bottom loading indicator will be displayed.
   ///
-  /// Check the [SBBCheckboxListItem.custom] constructor for a complete customization.
-  SBBCheckboxListItem({
+  /// Check the [SBBRadioListItem.custom] constructor for a complete customization.
+  SBBRadioListItem({
     Key? key,
-    required bool? value,
+    required T value,
+    required T? groupValue,
     required String label,
-    required ValueChanged<bool?>? onChanged,
-    bool tristate = false,
-    bool isLastElement = false,
+    required ValueChanged<T?>? onChanged,
     bool allowMultilineLabel = false,
     String? secondaryLabel,
+    bool isLastElement = false,
     IconData? leadingIcon,
     IconData? trailingIcon,
     VoidCallback? onCallToAction,
     bool isLoading = false,
-    String? checkboxSemanticLabel,
+    String? radioSemanticLabel,
   }) : this.custom(
           key: key,
           value: value,
+          groupValue: groupValue,
           label: label,
-          onChanged: onChanged,
-          tristate: tristate,
-          isLastElement: isLastElement,
           allowMultilineLabel: allowMultilineLabel,
           secondaryLabel: secondaryLabel,
+          onChanged: onChanged,
+          isLastElement: isLastElement,
           leadingIcon: leadingIcon,
           trailingWidget: _optionallyButtonedTrailingIcon(trailingIcon, onCallToAction, onChanged),
           isLoading: isLoading,
-          checkboxSemanticLabel: checkboxSemanticLabel,
+          radioSemanticLabel: radioSemanticLabel,
         );
 
-  /// Use this in combination with a [SBBGroup] to create a boxed variant of the [SBBCheckboxListItem].
+  /// Use this in combination with a [SBBGroup] to create a boxed variant of the [SBBRadioListItem].
   ///
   /// ```dart
   /// SBBGroup(
-  ///   child: SBBCheckboxListItem(
-  ///     value: _throwShotAway,
-  ///     onChanged: (bool newValue) {
-  ///       setState(() {
-  ///         _throwShotAway = newValue;
-  ///       });
-  ///     },
-  ///     label: 'Example',
+  ///   child: SBBRadioButtonListItem<SingingCharacter>(
+  ///   value: SingingCharacter.lafayette,
+  ///   groupValue: _character,
+  ///   onChanged: (SingingCharacter newValue) {
+  ///     setState(() {
+  ///       _character = newValue;
+  ///     });
+  ///   },
   ///   )
   /// )
   ///
   /// ```
   ///
-  SBBCheckboxListItem.boxed({
+  SBBRadioListItem.boxed({
     Key? key,
-    required bool? value,
+    required T value,
+    required T? groupValue,
     required String label,
-    required ValueChanged<bool?>? onChanged,
-    bool tristate = false,
+    required ValueChanged<T?>? onChanged,
     bool allowMultilineLabel = false,
     String? secondaryLabel,
     IconData? leadingIcon,
     IconData? trailingIcon,
     VoidCallback? onCallToAction,
     bool isLoading = false,
-    String? checkboxSemanticLabel,
+    String? radioSemanticLabel,
   }) : this.custom(
           key: key,
           value: value,
+          groupValue: groupValue,
           label: label,
-          onChanged: onChanged,
-          tristate: tristate,
           allowMultilineLabel: allowMultilineLabel,
           secondaryLabel: secondaryLabel,
+          onChanged: onChanged,
+          isLastElement: true,
           leadingIcon: leadingIcon,
           trailingWidget: _optionallyButtonedTrailingIcon(trailingIcon, onCallToAction, onChanged),
           isLoading: isLoading,
-          isLastElement: true,
-          checkboxSemanticLabel: checkboxSemanticLabel,
+          radioSemanticLabel: radioSemanticLabel,
         );
 
-  /// Full custom [SBBCheckboxListItem].
-  ///
-  /// When using [isBoxed], wrap this widget with a SBBGroup to achieve the desired outcome.
-  const SBBCheckboxListItem.custom({
+  const SBBRadioListItem.custom({
     super.key,
     required this.value,
+    required this.groupValue,
     required this.label,
+    required this.onChanged,
     this.allowMultilineLabel = false,
     this.secondaryLabel,
-    this.tristate = false,
-    required this.onChanged,
     this.isLastElement = false,
     this.leadingIcon,
     this.trailingWidget,
     this.isLoading = false,
-    this.checkboxSemanticLabel,
-  }) : assert(tristate || value != null);
+    this.radioSemanticLabel,
+  });
 
-  /// Whether this checkbox is checked.
-  ///
-  /// If [tristate] is false (the default), [value] must not be null.
-  /// If [tristate] is true, SBBCheckbox displays a dash when [value] is null.
-  final bool? value;
+  /// The value represented by this radio button.
+  final T value;
 
-  /// Called when the value of the checkbox should change.
+  /// The currently selected value for a group of radio buttons.
   ///
-  /// The checkbox passes the new value to the callback but does not actually
-  /// change state until the parent widget rebuilds the checkbox with the new
-  /// value.
+  /// This radio button is considered selected if its [value] matches the
+  /// [groupValue].
+  final T? groupValue;
+
+  /// Called when the user selects this [SBBRadioListItem].
   ///
-  /// If this callback is null, the checkbox will be displayed as disabled
-  /// and will not respond to input gestures.
+  /// The [SBBRadioListItem] passes [value] as a parameter to this callback. The list item
+  /// does not actually change state until the parent widget rebuilds the
+  /// list item with the new [groupValue].
   ///
-  /// When the checkbox is tapped, if [tristate] is false (the default) then
-  /// the [onChanged] callback will be applied to `!value`. If [tristate] is
-  /// true this callback cycle from false to true to null.
+  /// If null, the list item will be displayed as disabled.
+  ///
+  /// The provided callback will not be invoked if this list item is already
+  /// selected.
   ///
   /// The callback provided to [onChanged] should update the state of the parent
   /// [StatefulWidget] using the [State.setState] method, so that the parent
   /// gets rebuilt; for example:
   ///
   /// ```dart
-  /// SBBCheckboxListItem(
-  ///   value: _throwShotAway,
-  ///   onChanged: (bool newValue) {
+  /// SBBRadioButtonListItem<SingingCharacter>(
+  ///   value: SingingCharacter.lafayette,
+  ///   groupValue: _character,
+  ///   onChanged: (SingingCharacter newValue) {
   ///     setState(() {
-  ///       _throwShotAway = newValue;
+  ///       _character = newValue;
   ///     });
   ///   },
-  ///   label: 'Example',
   /// )
   /// ```
-  final ValueChanged<bool?>? onChanged;
-
-  /// If true the checkbox's [value] can be true, false, or null.
-  ///
-  /// [SBBCheckbox] displays a dash when its value is null.
-  ///
-  /// When a tri-state checkbox ([tristate] is true) is tapped, its [onChanged]
-  /// callback will be applied to true if the current value is false, to null if
-  /// value is true, and to false if value is null (i.e. it cycles through false
-  /// => true => null => false when tapped).
-  ///
-  /// If tristate is false (the default), [value] must not be null.
-  final bool tristate;
+  final ValueChanged<T?>? onChanged;
 
   /// The primary label on the tile.
   final String label;
@@ -208,76 +192,52 @@ class SBBCheckboxListItem extends StatelessWidget {
   /// Whether to display a BottomLoadingIndicator on the [SBBCheckboxListItem].
   final bool isLoading;
 
-  /// The semantic label for the checkbox that will be announced by screen readers.
+  /// The semantic label for the RadioButton that will be announced by screen readers.
   ///
   /// This is announced by assistive technologies (e.g TalkBack/VoiceOver).
   ///
   /// This label does not show in the UI.
-  final String? checkboxSemanticLabel;
+  final String? radioSemanticLabel;
 
   /// Whether [value] of this control can be changed by user interaction.
   ///
   /// The control is considered interactive if the [onChanged] callback is
   /// non-null. If the callback is null, then the control is disabled, and
-  /// non-interactive. A disabled checkbox, for example, is displayed using a
-  /// grey color and its value cannot be changed.
+  /// non-interactive.
   bool get _isInteractive => onChanged != null;
-
-  static Widget? _optionallyButtonedTrailingIcon(
-    IconData? trailingIcon,
-    VoidCallback? onCallToAction,
-    ValueChanged<bool?>? onChanged,
-  ) {
-    return trailingIcon == null
-        ? null
-        : onCallToAction == null
-            ? Padding(
-                padding: const EdgeInsets.only(top: _iconTopPadding, right: sbbDefaultSpacing),
-                child: Icon(trailingIcon, size: 24.0),
-              )
-            : Padding(
-                padding: const EdgeInsets.only(right: sbbDefaultSpacing * .5),
-                child: SBBIconButtonSmall(
-                  icon: trailingIcon,
-                  onPressed: onChanged != null ? onCallToAction : null,
-                ),
-              );
-  }
 
   @override
   Widget build(BuildContext context) {
-    final style = SBBControlStyles.of(context).checkbox;
-    final Color? resolvedBackgroundColor = _resolveBackgroundColor(style);
-
+    final style = SBBControlStyles.of(context).radioButton;
     return MergeSemantics(
       child: Material(
-        color: resolvedBackgroundColor,
+        color: _resolveBackgroundColor(style),
         child: Stack(
-          alignment: AlignmentDirectional.bottomStart,
+          alignment: Alignment.bottomCenter,
           children: [
             ConstrainedBox(
               constraints: const BoxConstraints(minHeight: _minListTileHeight),
               child: InkWell(
-                onTap: _isInteractive ? _handleTap : null,
+                onTap: _isInteractive ? () => onChanged!(value) : null,
                 splashColor: style?.listItem?.backgroundColorHighlighted,
                 focusColor: style?.listItem?.backgroundColorHighlighted,
                 highlightColor: SBBColors.transparent,
                 hoverColor: SBBColors.transparent,
                 child: Semantics(
                   enabled: _isInteractive,
-                  child: _checkboxBody(style),
+                  child: _radioTileBody(style),
                 ),
               ),
             ),
             if (!isLastElement) const Divider(),
-            if (isLoading) BottomLoadingIndicator()
+            if (isLoading) BottomLoadingIndicator(),
           ],
         ),
       ),
     );
   }
 
-  Widget _checkboxBody(SBBControlStyle? style) {
+  Widget _radioTileBody(SBBControlStyle? style) {
     final Color? resolvedIconColor = _resolveIconColor(style);
 
     return IconTheme.merge(
@@ -286,11 +246,11 @@ class SBBCheckboxListItem extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           const SizedBox(width: sbbDefaultSpacing),
-          _NonHittableCheckbox(
+          _NonHittableRadio<T>(
             value: value,
-            tristate: tristate,
+            groupValue: groupValue,
             onChanged: onChanged,
-            semanticLabel: checkboxSemanticLabel,
+            radioSemanticLabel: radioSemanticLabel,
           ),
           const SizedBox(width: sbbDefaultSpacing * 0.5),
           if (leadingIcon != null) _LeadingIcon(leadingIcon: leadingIcon),
@@ -309,26 +269,35 @@ class SBBCheckboxListItem extends StatelessWidget {
     );
   }
 
-  void _handleTap() {
-    if (!_isInteractive) {
-      return;
-    }
-    switch (value) {
-      case false:
-        onChanged!(true);
-      case true:
-        onChanged!(tristate ? null : false);
-      case null:
-        onChanged!(false);
-    }
-  }
-
   Color? _resolveBackgroundColor(SBBControlStyle? style) {
     return _isInteractive ? style?.listItem?.backgroundColor : style?.listItem?.backgroundColorDisabled;
   }
 
   Color? _resolveIconColor(SBBControlStyle? style) {
     return _isInteractive ? style?.listItem?.iconColor : style?.listItem?.iconColorDisabled;
+  }
+
+  static Widget? _optionallyButtonedTrailingIcon<T>(
+    IconData? trailingIcon,
+    VoidCallback? onCallToAction,
+    ValueChanged<T?>? onChanged,
+  ) {
+    if (trailingIcon == null) return null;
+
+    if (onCallToAction == null) {
+      return Padding(
+        padding: const EdgeInsets.only(top: _iconTopPadding, right: sbbDefaultSpacing),
+        child: Icon(trailingIcon, size: 24.0),
+      );
+    } else {
+      return Padding(
+        padding: const EdgeInsets.only(right: sbbDefaultSpacing * .5),
+        child: SBBIconButtonSmall(
+          icon: trailingIcon,
+          onPressed: onChanged != null ? onCallToAction : null,
+        ),
+      );
+    }
   }
 }
 
@@ -395,18 +364,18 @@ class _LeadingIcon extends StatelessWidget {
   }
 }
 
-class _NonHittableCheckbox extends StatelessWidget {
-  const _NonHittableCheckbox({
+class _NonHittableRadio<T> extends StatelessWidget {
+  const _NonHittableRadio({
     required this.value,
-    required this.tristate,
+    required this.groupValue,
     required this.onChanged,
-    this.semanticLabel,
+    this.radioSemanticLabel,
   });
 
-  final bool? value;
-  final bool tristate;
-  final ValueChanged<bool?>? onChanged;
-  final String? semanticLabel;
+  final T value;
+  final T? groupValue;
+  final ValueChanged<T?>? onChanged;
+  final String? radioSemanticLabel;
 
   @override
   Widget build(BuildContext context) {
@@ -416,12 +385,12 @@ class _NonHittableCheckbox extends StatelessWidget {
           padding: const EdgeInsetsDirectional.only(
             top: sbbDefaultSpacing * 0.75,
           ),
-          child: SBBCheckbox(
+          child: SBBRadio(
             value: value,
-            tristate: tristate,
+            groupValue: groupValue,
             onChanged: onChanged,
             padding: EdgeInsets.zero,
-            semanticLabel: semanticLabel,
+            semanticLabel: radioSemanticLabel,
           ),
         ),
       ),